+++ tests/LoggingTest.php
@@ -136,6 +136,25 @@ class LoggingTest extends TestCase
 
     public function testProcessorTap()
     {
+        $expectedForwardedFor = '192.168.1.1, 10.0.0.1';
+        $expectedHttpMethod = 'GET';
+        $expectedIp = '192.168.1.1';
+        $expectedReferrer = 'https://duckduckgo.com/';
+        $expectedServer = 'duckduckgo.com';
+        $expectedUniqueId = '03ef5820-228c-4344-b43a-c9c8bc36e3bf';
+        $expectedUrl = '/home';
+        $expectedUserAgent = 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/115.0.0.0'
+            . ' Safari/537.36';
+
+        $_SERVER['HTTP_REFERER'] = $expectedReferrer;
+        $_SERVER['HTTP_USER_AGENT'] = $expectedUserAgent;
+        $_SERVER['HTTP_X_FORWARDED_FOR'] = $expectedForwardedFor;
+        $_SERVER['HTTP_X_AMZN_TRACE_ID'] = $expectedUniqueId;
+        $_SERVER['REMOTE_ADDR'] = $expectedIp;
+        $_SERVER['REQUEST_METHOD'] = $expectedHttpMethod;
+        $_SERVER['REQUEST_URI'] = $expectedUrl;
+        $_SERVER['SERVER_NAME'] = $expectedServer;
+
         // configure a test logger
         $handler = new TestHandler();
         $logger = new Logger('testing', [$handler]);
@@ -148,9 +167,26 @@ class LoggingTest extends TestCase
 
         // assert the log value contains the extra keys
         $record = $handler->getRecords()[0]['extra'];
+
+        $this->assertArrayHasKey('forwarded_for', $record);
+        $this->assertEquals($record['forwarded_for'], $expectedForwardedFor);
+        $this->assertArrayHasKey('http_method', $record);
+        $this->assertEquals($record['http_method'], $expectedHttpMethod);
+        $this->assertArrayHasKey('ip', $record);
+        $this->assertEquals($record['ip'], $expectedIp);
         $this->assertArrayHasKey('memory_peak_usage', $record);
         $this->assertArrayHasKey('memory_usage', $record);
+        $this->assertArrayHasKey('referrer', $record);
+        $this->assertEquals($record['referrer'], $expectedReferrer);
+        $this->assertArrayHasKey('server', $record);
+        $this->assertEquals($record['server'], $expectedServer);
         $this->assertArrayHasKey('uid', $record);
+        $this->assertArrayHasKey('unique_id', $record);
+        $this->assertEquals($record['unique_id'], $expectedUniqueId);
+        $this->assertArrayHasKey('url', $record);
+        $this->assertEquals($record['url'], $expectedUrl);
+        $this->assertArrayHasKey('user_agent', $record);
+        $this->assertEquals($record['user_agent'], $expectedUserAgent);
     }
 
     public function testLogstashSingleChannel()